KNOWN_BUGS: restore text regarding #2101.
This was added earlier but appears to have been removed accidentally.

AFAICT this is very much still an issue.

-----

I say "accidentally" because the text seems to have harmlessly snuck
into [1] (which makes no mention of it).  [1] was later reverted for
unspecified reasons in [2], presumably because the mentioned issue was
fixed or invalid.

[1] fb4f15e8ff092799554dbaf6f6fb045aa7cf8f0e
[2] 7b92cd7be8ed38e32552e8d2c0a1a978aa0bfe8e

Closes #2618