CURLOPT_ADDRESS_SCOPE: fix range check and more
Commit 599b537 fixed most of the confusing issues between scope id and
scope however f32f179 added bad limits checking assuming that the scope
is being set and not the scope id.

I have fixed the documentation so it all refers to scope ids.

In addition Curl_if2ip refered to the scope id as remote_scope_id which
is incorrect, so I renamed it to local_scope_id.

Adjusted-by: Daniel Stenberg

Closes #3655
Closes #3765
Fixes #3713