winbuild: Add manifest to curl.exe for proper OS version detection
This is a small fix to commit 7c145e329bacbb9ab90528f7bebcf65015890670
in pull request #1221. That commit added the CURL_EMBED_MANIFEST flag to
CURL_RC_FLAGS. However, later in the file CURL_RC_FLAGS is
overwritten. The fix is to append values to CURL_RC_FLAGS instead of
overwriting

Closes #4399