cmake: set HAVE_GETADDRINFO_THREADSAFE on Windows
autotools enables this configuration option unconditionally for Windows
[^1]. Do the same in CMake.

The above will make this work for all reasonably recent environments.
The logic present in `lib/config-win32.h` [^2] has the following
exceptions which we did not cover in this CMake update:

- Builds targeting Windows 2000 and earlier
- MS Visual C++ 5.0 (1997) and earlier

Also make sure to disable this feature when `HAVE_GETADDRINFO` isn't
set, to avoid a broken build. We might want to handle that in the C
sources in a future commit.

[^1]: https://github.com/curl/curl/blob/b102fc336026a1c128d2e503348f4c2e84b15849/m4/curl-functions.m4#L2067-L2070

[^2]: https://github.com/curl/curl/blob/b102fc336026a1c128d2e503348f4c2e84b15849/lib/config-win32.h#L511-L528

Closes #9727